requirements.txt
feedparser
newspaper4k
semantic-kernel
openai
python-dotenv
lxml_html_clean